+++ ta4j-core/src/main/java/org/ta4j/core/indicators/AccelerationDecelerationIndicator.java
@@ -63,7 +63,7 @@ public class AccelerationDecelerationIndicator extends CachedIndicator<Num> {
     }
 
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return 0;
     }
 }