+++ ta4j-core/src/main/java/org/ta4j/core/indicators/AwesomeOscillatorIndicator.java
@@ -86,7 +86,7 @@ public class AwesomeOscillatorIndicator extends CachedIndicator<Num> {
     }
 
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return 0;
     }
 }