+++ ta4j-core/src/main/java/org/ta4j/core/indicators/adx/DXIndicator.java
@@ -65,7 +65,7 @@ public class DXIndicator extends CachedIndicator<Num> {
     }
 
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return barCount;
     }