+++ ta4j-core/src/main/java/org/ta4j/core/indicators/candles/LowerShadowIndicator.java
@@ -65,7 +65,7 @@ public class LowerShadowIndicator extends CachedIndicator<Num> {
     }
 
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return 0;
     }
 }