+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/LowestValueIndicator.java
@@ -70,7 +70,7 @@ public class LowestValueIndicator extends CachedIndicator<Num> {
 
     /** @return {@link #barCount} */
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return barCount;
     }