+++ ta4j-core/src/main/java/org/ta4j/core/indicators/volume/AccumulationDistributionIndicator.java
@@ -61,7 +61,7 @@ public class AccumulationDistributionIndicator extends RecursiveCachedIndicator<
     }
 
     @Override
-    public int getUnstableBars() {
+    public int getCountOfUnstableBars() {
         return 0;
     }
 }